Add files via upload
Challenge Project: Company Home Page with Flexbox
You’re going to build a company homepage for a real or imagined company. You’ll get to decide exactly what content to include for this company, but you should at least include:

A title and logo or splash image.
A mission statement or brief description of the company.
A list or set of images and titles representing the product or products of the company.
A section describing some of the company’s employees or teammates.
If you want, you can tour our example site for inspiration or to see how we used flexbox. Your site will probably look very different from ours, and that’s great!